requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
transformers>=4.42.0
torch>=2.2.0
pytest>=8.2.0
python-docx>=1.1.2
PyMuPDF>=1.24.0
httpx>=0.27.0
